+++ lib/src/rules_factory.dart
@@ -13,6 +13,7 @@ import 'rules/prefer_conditional_expressions.dart';
 import 'rules/prefer_intl_name.dart';
 import 'rules/prefer_on_push_cd_strategy.dart';
 import 'rules/prefer_trailing_comma_for_collection.dart';
+import 'rules/provide_correct_intl_args.dart';
 
 final _implementedRules = <String, BaseRule Function(Map<String, Object>)>{
   AvoidPreserveWhitespaceFalseRule.ruleId: (config) =>
@@ -34,6 +35,8 @@ final _implementedRules = <String, BaseRule Function(Map<String, Object>)>{
       NoObjectDeclarationRule(config: config),
   PreferConditionalExpressions.ruleId: (config) =>
       PreferConditionalExpressions(config: config),
+  ProvideCorrectIntlArgsRule.ruleId: (config) =>
+      ProvideCorrectIntlArgsRule(config: config),
   PreferIntlNameRule.ruleId: (config) => PreferIntlNameRule(config: config),
   PreferTrailingCommaForCollectionRule.ruleId: (config) =>
       PreferTrailingCommaForCollectionRule(config: config),